+++ src/lib/index.ts
@@ -1,4 +1,5 @@
 export * from './core';
+export * from './tooltip';
 export * from './axis';
 export * from './bars';
 export * from './points';